ai-rag-pdf-assistant/backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
faiss-cpu
numpy
sentence-transformers
pypdf
python-multipart